// src/world/regions/generated/indian.mask.ts
export const CELL_W = 8;
export const CELL_H = 11;
// ...
export const ANCHORS: Record<string, { x: number; y: number }> = {
  colombo: { x: 473, y: 330 },
  c_dock: { x: 456, y: 339 },
  c_spawn: { x: 438, y: 365 },
  school_i_1: { x: 243, y: 248 },
  school_i_2: { x: 307, y: 346 },
  school_s_1: { x: 250, y: 923 },
  school_s_2: { x: 429, y: 972 },
  label_malacca: { x: 712, y: 390 },
};